Wait for volume event while changing volume
This helps a UI glitch introduced when converting to AsyncSockets that
would cause the master volume slider to bounce backwards a few steps
when holding down the keyboard shortcut for incrementing or decrementing
volume. The glitch was caused by commands sent getting ahead of the
responses being reflected by the bound slider control.